Bug fix in LAUNCH_REQUEST/LAUNCH_RESPONSE protocol.
JavaSockets is expecting a 24 byte return message, the
native sockets launcher was only sending 20 bytes (thus
leading to a hang in JavaSockets which waited forever
for 4 bytes that were never going to arrive).

+++ x10.runtime/x10rt/sockets/Launcher.cc
@@ -1018,7 +1018,7 @@ int Launcher::handleControlMessage(int fd)
 				m.to = m.from;
 				m.from = _myproc;
 				m.type = LAUNCH_RESPONSE;
-				int zero = 0;
+				int64_t zero = 0;
 				m.datalen = sizeof(zero);
 				#ifdef DEBUG
 					fprintf(stderr, "Launcher %d responding to an unsupported launch request \n", _myproc);
